Move the base request/response primatives to the p2p module
