+++ p2p/exchange/validator.py
@@ -1,0 +1,5 @@
+from typing import Any
+
+
+def noop_payload_validator(request: Any, response: Any) -> None:
+    pass
